Refactor RE
